update coding syntax to make pages consistent
