+++ client/components/index.js
@@ -3,14 +3,14 @@
  * This way, we can import all of our components from the same place, rather than
  * having to figure out which file they belong to!
  */
+export { default as AboutUs } from "./BeforeLoginPages/AboutUs";
+export { default as HomePage } from "./BeforeLoginPages/HomePage";
+export { default as HowItWorks } from "./BeforeLoginPages/HowItWorks";
+export { default as Login } from "./BeforeLoginPages/Login";
+export { default as SignUp } from "./BeforeLoginPages/SignUp";
 export { default as Navbar } from "./navbar";
 export { default as UserHome } from "./user-home";
 //export { Login, Signup } from "./auth-form";
-export { default as HomePage } from "./HomePage";
-export { default as SignUp } from "./SignUp";
-export { default as Login } from "./Login";
-export { default as AboutUs } from "./AboutUs";
-export { default as HowItWorks } from "./HowItWorks";
 export { default as DataMapHousing } from "./DataMapHousing";
 export { default as UserTrend } from "./UserTrend";
 export { default as Transactions } from "./Transactions";
@@ -18,7 +18,7 @@ export { default as AccountTable } from "./AccountTable";
 export { default as Profile } from "./Profile";
 export { default as Budget } from "./Budget";
 export {
-  default as BudgetHorizontalBarChart
+  default as BudgetHorizontalBarChart,
 } from "./BudgetHorizontalBarChart";
 export { default as BudgetProgress } from "./BudgetProgress";
 export { default as BudgetModal } from "./BudgetModal";
@@ -29,12 +29,12 @@ export { default as RetirementResult } from "./RetirementResult";
 export { default as UserCredit } from "./UserCredit";
 export { default as Goals } from "./Goals";
 export {
-  default as SaveForEmergencyMainPage
+  default as SaveForEmergencyMainPage,
 } from "./SaveForEmergencyMainPage";
 export {
-  default as DataMapApartmentsMMRentChange
+  default as DataMapApartmentsMMRentChange,
 } from "./DataMapApartmentsMMRentChange";
 export {
-  default as DataMapApartmentYYRentChange
+  default as DataMapApartmentYYRentChange,
 } from "./DataMapApartmentYYRentChange";
 export { default as House } from "./house";
